requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==2.0
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
jsonfield==2.0.2
mysqlclient==2.0.3
Pillow==5.0.0
PyMySQL==1.0.2
pytz==2017.3
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.10.0.0
wincertstore==0.2
